Add setting LLVM_LINK_LLVM_DYLIB and updated "help"
With LLVM_LINK_LLVM_DYLIB=ON will cause clang to default the compiled
programs to using dynamic linking. On linux you may need to set the
LD_LIBRARY_PATH enviornment variable to your LLVM_INSTALL_PATH so the
libLLVM-X.so is found when running the app. This may not be required if
the -L<LLVM_INSTALL_PATH> is passed to clang.